Create a custom hook for windowIsLarge logic
